pyproject.toml
[project]
name = "kumabot"
version = "25.23.0"
description = "一种舞萌DX机器人"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.14",
    "nanoid>=2.0.0",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot2[fastapi]>=2.4.2",
    "numpy>=2.3.2",
    "openai[aiohttp]>=1.98.0",
    "orjson>=3.11.1",
    "pillow>=11.3.0",
    "pyhocon>=0.3.61",
    "pykakasi>=2.3.0",
    "rapidfuzz>=3.13.0",
    "soundfile>=0.13.1",
    "sqlalchemy[postgresql-asyncpg]>=2.0.41",
    "xxhash>=3.5.0",
]
